+++ backend/controllers/messages.js
@@ -1,10 +1,10 @@
-const MessagesRouter = require("express").Router();
+const MessagesRouter = require('express').Router();
 
-const postgreSql = require("../db.js");
-const { checkChatAccess, checkMessageAccess } = require("../utils/middleware.js");
+const postgreSql = require('../db.js');
+const { checkChatAccess, checkMessageAccess, fieldWhiteList } = require('../utils/middleware.js');
 // const config = require('../utils/config.js');
 
-MessagesRouter.get("/", async (request, response) => {
+MessagesRouter.get('/', async (request, response) => {
   try {
     const messages = await postgreSql`
       SELECT public_id, message
@@ -14,111 +14,101 @@ MessagesRouter.get("/", async (request, response) => {
     response.json(messages);
   } catch (error) {
     console.log(error);
-    response.status(500).json({ message: "Internal server error" });
+    response.status(500).json({ message: 'Internal server error' });
   }
 });
-MessagesRouter.get(
-  "/message/:message_public_id",
-  checkMessageAccess,
-  async (request, response) => {
-    try {
-      const [message] = await postgreSql`
-      SELECT * FROM messages
+MessagesRouter.get('/message/:message_public_id', async (request, response) => {
+  try {
+    const [message] = await postgreSql`
+      SELECT message, created_at, edited_at, public_id
+      FROM messages
       WHERE public_id = ${request.params.message_public_id}
     `;
 
-      if (!message) {
-        response.status(404).json({ message: "Message not found" });
-      }
-
-      response.json(message);
-    } catch (error) {
-      console.log(error);
-      response.status(500).json({ message: "Internal server error" });
-    }
-  },
-);
-MessagesRouter.get(
-  "/user/:user_public_id",
-  checkMessageAccess,
-  async (request, response) => {
-    if (request.userRoleInChat !== "owner") {
-      return response
-        .status(403)
-        .json({ message: "Only owners can delete the chat" });
+    if (!message) {
+      response.status(404).json({ message: 'Message not found' });
     }
 
-    try {
-      const messages = await postgreSql`
-      WITH user_data AS (
-          SELECT id FROM chat.users WHERE public_id = ${request.params.user_public_id}
-      ),
-      last_messages AS (
-          SELECT DISTINCT ON (m.chat_id) 
-              m.chat_id,
-              m.message,
-              m.created_at,
-              u.public_id AS sender_public_id
-          FROM chat.messages m
-          JOIN chat.users u ON m.sender_id = u.id
-          WHERE m.chat_id IN (SELECT chat_id FROM chat.chats_members WHERE user_id = (SELECT id FROM user_data))
-          ORDER BY m.chat_id, m.created_at DESC
-      )
-      SELECT 
-          c.public_id AS chat_public_id,
-          CASE 
-              WHEN c.type = 'private' THEN (
-                  SELECT u2.name 
-                  FROM chat.chats_members cm2
-                  JOIN chat.users u2 ON cm2.user_id = u2.id
-                  WHERE cm2.chat_id = c.id AND cm2.user_id != (SELECT id FROM user_data)
-                  LIMIT 1
-              )
-              ELSE c.name 
-          END AS display_name,
-          c.type AS chat_type,
-          lm.message AS last_message_text,
-          lm.created_at AS last_message_time,
-          lm.sender_public_id AS last_message_sender_id
-      FROM chat.chats_members cm
-      JOIN chat.chats c ON cm.chat_id = c.id
-      LEFT JOIN last_messages lm ON c.id = lm.chat_id
-      WHERE cm.user_id = (SELECT id FROM user_data)
-      ORDER BY lm.created_at DESC NULLS LAST;
-    `;
+    response.json(message);
+  } catch (error) {
+    console.log(error);
+    response.status(500).json({ message: 'Internal server error' });
+  }
+});
+// MessagesRouter.get('/user/:user_public_id', checkChatAccess, async (request, response) => {
+//   if (request.userRoleInChat === 'undefined') {
+//     return response.status(403).json({
+//       message: 'Only users with access can see messages of this chat',
+//     });
+//   }
 
-      if (!messages) {
-        response.status(404).json({ message: "Message not found" });
-      }
+//   try {
+//     const messages = await postgreSql`
+//       WITH user_data AS (
+//           SELECT id FROM chat.users WHERE public_id = ${request.params.user_public_id}
+//       ),
+//       last_messages AS (
+//           SELECT DISTINCT ON (m.chat_id) 
+//               m.chat_id,
+//               m.message,
+//               m.created_at,
+//               u.public_id AS sender_public_id
+//           FROM chat.messages m
+//           JOIN chat.users u ON m.sender_id = u.id
+//           WHERE m.chat_id IN (SELECT chat_id FROM chat.chats_members WHERE user_id = (SELECT id FROM user_data))
+//           ORDER BY m.chat_id, m.created_at DESC
+//       )
+//       SELECT 
+//           c.public_id AS chat_public_id,
+//           CASE 
+//               WHEN c.type = 'private' THEN (
+//                   SELECT u2.name 
+//                   FROM chat.chats_members cm2
+//                   JOIN chat.users u2 ON cm2.user_id = u2.id
+//                   WHERE cm2.chat_id = c.id AND cm2.user_id != (SELECT id FROM user_data)
+//                   LIMIT 1
+//               )
+//               ELSE c.name 
+//           END AS display_name,
+//           c.type AS chat_type,
+//           lm.message AS last_message_text,
+//           lm.created_at AS last_message_time,
+//           lm.sender_public_id AS last_message_sender_id
+//       FROM chat.chats_members cm
+//       JOIN chat.chats c ON cm.chat_id = c.id
+//       LEFT JOIN last_messages lm ON c.id = lm.chat_id
+//       WHERE cm.user_id = (SELECT id FROM user_data)
+//       ORDER BY lm.created_at DESC NULLS LAST;
+//     `;
 
-      response.json(messages);
-    } catch (error) {
-      console.log(error);
-      response.status(500).json({ message: "Internal server error" });
-    }
-  },
-);
-MessagesRouter.get(
-  "/chat/:chat_public_id",
-  checkChatAccess,
-  async (request, response) => {
-    if (request.userRoleInChat !== "owner") {
-      return response
-        .status(403)
-        .json({ message: "Only owners can delete the chat" });
-    }
+//     if (!messages) {
+//       response.status(404).json({ message: 'Message not found' });
+//     }
 
-    try {
-      const { chat_public_id } = request.params;
-      const user_public_id = request.user.public_id;
+//     response.json(messages);
+//   } catch (error) {
+//     console.log(error);
+//     response.status(500).json({ message: 'Internal server error' });
+//   }
+// });
+MessagesRouter.get('/chat/:chat_public_id', checkChatAccess, async (request, response) => {
+  if (request.userRoleInChat === 'undefined') {
+    return response.status(403).json({
+      message: 'Only users with access can see messages of this chat',
+    });
+  }
 
-      const messages = await postgreSql`
+  try {
+    const chatId = request.chatInternalId;
+    const user_public_id = request.user.public_id;
+
+    const messages = await postgreSql`
       WITH updated_member AS (
     
           UPDATE chat.chats_members
           SET last_read_at = CURRENT_TIMESTAMP
           WHERE user_id = (SELECT id FROM chat.users WHERE public_id = ${user_public_id})
-            AND chat_id = (SELECT id FROM chat.chats WHERE public_id = ${chat_public_id})
+            AND chat_id = (SELECT id FROM chat.chats WHERE id = ${chatId})
           RETURNING chat_id
       )
 
@@ -146,27 +136,26 @@ MessagesRouter.get(
       ORDER BY m.created_at ASC;
     `;
 
-      response.json(messages);
-    } catch (error) {
-      response.status(500).json({ error: "Failed to fetch messages" });
-    }
-  },
-);
+    response.json(messages);
+  } catch (error) {
+    response.status(500).json({ error: 'Failed to fetch messages' });
+  }
+});
 
-MessagesRouter.post("/", checkChatAccess, async (request, response) => {
-  if (request.userRoleInChat !== "owner") {
-    return response
-      .status(403)
-      .json({ message: "Only owners can delete the chat" });
+MessagesRouter.post('/', checkChatAccess, async (request, response) => {
+  if (request.userRoleInChat === 'undefined') {
+    return response.status(403).json({
+      message: 'Only users with access can add messages to this chat',
+    });
   }
 
   try {
-    const { message, chat_public_id, additionals } = request.body;
+    const { message, additionals } = request.body;
+    const chatId = request.chatInternalId;
     const sender_id = request.user.id;
 
     const result = await postgreSql.begin(async (sql) => {
-      const [chat] =
-        await sql`SELECT id FROM chat.chats WHERE public_id = ${chat_public_id}`;
+      const [chat] = await sql`SELECT id FROM chat.chats WHERE public_id = ${chatId}`;
 
       const [newMessage] = await sql`
         INSERT INTO chat.messages (chat_id, sender_id, message)
@@ -186,13 +175,15 @@ MessagesRouter.post("/", checkChatAccess, async (request, response) => {
     response.status(201).json(result);
   } catch (error) {
     console.error(error);
-    response.status(500).json({ message: "Internal server error" });
+    response.status(500).json({ message: 'Internal server error' });
   }
 });
 
 MessagesRouter.put(
-  "/:public_id",
+  '/:public_id',
   checkChatAccess,
+  checkMessageAccess,
+  fieldWhiteList,
   async (request, response) => {
     try {
       const { fieldData } = request.body;
@@ -208,15 +199,22 @@ MessagesRouter.put(
       response.status(201).json(updatedMessage);
     } catch (error) {
       console.log(error);
-      response.status(500).json({ message: "Internal server error" });
+      response.status(500).json({ message: 'Internal server error' });
     }
   },
 );
 
 MessagesRouter.delete(
-  "/message/:public_id",
+  '/message/:public_id',
   checkChatAccess,
+  checkMessageAccess,
   async (request, response) => {
+    if (request.userRoleInChat !== 'owner' || request.userRoleInChat !== 'admin') {
+      return response.status(403).json({
+        message: 'Only owners or admins can delete messages in this chat',
+      });
+    }
+
     try {
       const [deletedMessage] = await postgreSql`
       DELETE CASCADE from messages
@@ -227,29 +225,27 @@ MessagesRouter.delete(
       response.status(201).json(deletedMessage);
     } catch (error) {
       console.log(error);
-      response.status(500).json({ message: "Internal server error" });
+      response.status(500).json({ message: 'Internal server error' });
     }
   },
 );
-MessagesRouter.delete(
-  "/chat/:chat_public_id",
-  checkChatAccess,
-  async (request, response) => {
-    if (request.userRoleInChat !== "owner") {
-      return response
-        .status(403)
-        .json({ message: "Only owners can delete the chat" });
-    }
-    try {
-      const deletedMessages = await postgreSql`
+MessagesRouter.delete('/chat/:chat_public_id', checkChatAccess, async (request, response) => {
+  if (request.userRoleInChat !== 'owner' || request.userRoleInChat !== 'admin') {
+    return response.status(403).json({
+      message: 'Only owners or admins can delete messages in this chat',
+    });
+  }
+  try {
+    const chatId = request.chatInternalId;
+
+    const deletedMessages = await postgreSql`
       DELETE FROM chat.messages
-      WHERE chat_id = (SELECT id FROM chat.chats WHERE public_id = ${request.params.chat_public_id})
+      WHERE chat_id = (SELECT id FROM chat.chats WHERE id = ${chatId})
       RETURNING *
     `;
-      response.json({ count: deletedMessages.length });
-    } catch (error) {
-      console.log(error);
-      response.status(500).json({ message: "Internal server error" });
-    }
-  },
-);
+    response.json({ count: deletedMessages.length });
+  } catch (error) {
+    console.log(error);
+    response.status(500).json({ message: 'Internal server error' });
+  }
+});
